Fixed some error handlin around Advisor creation

+++ app/models/AdvisorForm.scala
@@ -1,6 +1,6 @@
 package models
 
-case class AdvisorForm(sin: String, name: String, workPhoneNumber: Int, homePhoneNumber: Int,
+case class AdvisorForm(sin: String, name: String, workPhoneNumber: Long, homePhoneNumber: Long,
                   streetAddress: String, city: String, postalcode: String, reportsTo: String) {
 
 }
